Long not long

+++ src/main/java/hu/domokosdaniel56/cosy/domain/User.java
@@ -11,7 +11,7 @@ public class User {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
+	private Long id;
 	private String username;
 	private String password;
 	
@@ -19,17 +19,17 @@ public class User {
 		
 	}
 	
-	public User(int id, String username, String password) {
+	public User(Long id, String username, String password) {
 		this.id = id;
 		this.username = username;
 		this.password = password;
 	}
 
-	public long getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
